change to double

+++ cpp/cudnn/include/helper.h
@@ -202,9 +202,9 @@ public:
     };
     cudnnHandle_t cudnn() { return _cudnn_handle; };
 
-    const float one = 1.f;
-    const float zero = 0.f;
-    const float minus_one = -1.f;
+    const double one = 1.f;
+    const double zero = 0.f;
+    const double minus_one = -1.f;
 
 private:
     cublasHandle_t _cublas_handle;
