+++ cpp/cudnn/include/layers/softmax.cuh
@@ -12,12 +12,12 @@ public:
     explicit Softmax(std::string name);
     ~Softmax() override = default;
 
-    Tensor<float> *forward(Tensor<float> *input) override;
-    Tensor<float> *backward(Tensor<float> *grad_input) override;
+    Tensor<double> *forward(Tensor<double> *input) override;
+    Tensor<double> *backward(Tensor<double> *grad_input) override;
 
 protected:
-    void fwd_initialize(Tensor<float> *input) override;
-    void bwd_initialize(Tensor<float> *grad_output) override;
+    void fwd_initialize(Tensor<double> *input) override;
+    void bwd_initialize(Tensor<double> *grad_output) override;
 };
 
 #endif // PROJECTNAME_SOFTMAX_CUH
