+++ cpp/cudnn/include/resnet.cuh
@@ -244,7 +244,7 @@ public:
                    relu11, relu12, relu13, relu14, relu15, relu16, pool1,  pool2,  dense1, softmax1,
                    add1,   add2,   add3,   add4,   add5,   add6,   add7,   add8};
     }
-    Tensor<float> *forward(Tensor<float> *input) override {
+    Tensor<double> *forward(Tensor<double> *input) override {
         auto output = this->conv1->forward(input);
         output = this->bn1->forward(output);
         output = this->relu1->forward(output);
@@ -256,7 +256,7 @@ public:
         if (identity1) {
             identity1->download(*output);
         } else {
-            identity1 = new Tensor<float>(*output);
+            identity1 = new Tensor<double>(*output);
         }
         output = this->conv2->forward(output);
         output = this->bn2->forward(output);
@@ -293,7 +293,7 @@ public:
         if (identity2) {
             identity2->download(*output);
         } else {
-            identity2 = new Tensor<float>(*output);
+            identity2 = new Tensor<double>(*output);
         }
         output = this->conv12->forward(output);
         output = this->bn12->forward(output);
@@ -339,7 +339,7 @@ public:
         if (identity3) {
             identity3->download(*output);
         } else {
-            identity3 = new Tensor<float>(*output);
+            identity3 = new Tensor<double>(*output);
         }
         output = this->conv25->forward(output);
         output = this->bn25->forward(output);
@@ -395,7 +395,7 @@ public:
         if (identity4) {
             identity4->download(*output);
         } else {
-            identity4 = new Tensor<float>(*output);
+            identity4 = new Tensor<double>(*output);
         }
         output = this->conv41->forward(output);
         output = this->bn41->forward(output);
@@ -434,7 +434,7 @@ public:
 
         return output;
     }
-    void backward(Tensor<float> *target) override {
+    void backward(Tensor<double> *target) override {
         auto gradient = this->softmax1->backward(target);
         gradient = this->dense1->backward(gradient);
         gradient = this->pool2->backward(gradient);
@@ -461,7 +461,7 @@ public:
         if (orig_grad1) {
             orig_grad1->download(*gradient);
         } else {
-            orig_grad1 = new Tensor<float>(*gradient);
+            orig_grad1 = new Tensor<double>(*gradient);
         }
         auto downsample_grad = this->bn44->backward(gradient);
         downsample_grad = this->conv44->backward(downsample_grad);
@@ -516,7 +516,7 @@ public:
         if (orig_grad2) {
             orig_grad2->download(*gradient);
         } else {
-            orig_grad2 = new Tensor<float>(*gradient);
+            orig_grad2 = new Tensor<double>(*gradient);
         }
         downsample_grad = this->bn28->backward(gradient);
         downsample_grad = this->conv28->backward(downsample_grad);
@@ -562,7 +562,7 @@ public:
         if (orig_grad3) {
             orig_grad3->download(*gradient);
         } else {
-            orig_grad3 = new Tensor<float>(*gradient);
+            orig_grad3 = new Tensor<double>(*gradient);
         }
         downsample_grad = this->bn15->backward(orig_grad3);
         downsample_grad = this->conv15->backward(downsample_grad);
@@ -599,7 +599,7 @@ public:
         if (orig_grad4) {
             orig_grad4->download(*gradient);
         } else {
-            orig_grad4 = new Tensor<float>(*gradient);
+            orig_grad4 = new Tensor<double>(*gradient);
         }
         downsample_grad = this->bn5->backward(gradient);
         downsample_grad = this->conv5->backward(downsample_grad);
@@ -751,14 +751,14 @@ private:
     Addition *add7;
     Addition *add8;
 
-    Tensor<float> *identity1 = nullptr;
-    Tensor<float> *identity2 = nullptr;
-    Tensor<float> *identity3 = nullptr;
-    Tensor<float> *identity4 = nullptr;
-    Tensor<float> *orig_grad1 = nullptr;
-    Tensor<float> *orig_grad2 = nullptr;
-    Tensor<float> *orig_grad3 = nullptr;
-    Tensor<float> *orig_grad4 = nullptr;
+    Tensor<double> *identity1 = nullptr;
+    Tensor<double> *identity2 = nullptr;
+    Tensor<double> *identity3 = nullptr;
+    Tensor<double> *identity4 = nullptr;
+    Tensor<double> *orig_grad1 = nullptr;
+    Tensor<double> *orig_grad2 = nullptr;
+    Tensor<double> *orig_grad3 = nullptr;
+    Tensor<double> *orig_grad4 = nullptr;
 };
 
 Network *make_resnet50() { return new ResNet50(); }
